BulletinBoard/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
confusable-homoglyphs==3.2.0
Django==4.0.5
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-filter==22.1
django-js-asset==2.0.0
django-registration==3.3
kombu==5.2.4
prompt-toolkit==3.0.30
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
